SignalShapingICARUSService split
SignalShapingICARUSService source was split into an implementation
(source and header) and a registration (service macros) part.
The header and implementation have lost the `_service` in the name,
so code including the header needs to be updated.
So now the idea is that:
 * icaruscode::TPC_Utilities_SignalShapingICARUSService_service contains
   `SignalShapingICARUSService.cc` and it is what user code should link to
 * icaruscode::TPC_Utilities_SignalShapingICARUSService_service_reg contains
   `SignalShapingICARUSService_service.cc` with the service "registration"
   macros; we don't bother with it, it's framework business
 * `icaruscode::TPC_Utilities` actually "contains" `SignalShapingICARUSService.h`;
   this is needed for the compiler to find the header, even if no library code
   is generated out of it. However, `icaruscode::TPC_Utilities_SignalShapingICARUSService_service`
   has a PUBLIC dependency on `icaruscode::TPC_Utilities`, so linking to the former
   enables the header discovery.

Not sure if this is a good strategy, but so far it seems to work.

Note that this naming pattern is the one expected by
`cet_build_plugin()` in `CMakeLists.txt`.
