+++ icaruscode/Decode/DecoderTools/FakeParticle_tool.cc
@@ -21,7 +21,7 @@
 #include "lardata/DetectorInfoServices/DetectorPropertiesService.h"
 
 // ICARUS package includes
-#include "icaruscode/TPC/Utilities/SignalShapingICARUSService_service.h"
+#include "icaruscode/TPC/Utilities/SignalShapingICARUSService.h"
 #include "icaruscode/Decode/DecoderTools/IFakeParticle.h"
 
 #include "icarus_signal_processing/Filters/ICARUSFFT.h"
