+++ icaruscode/TPC/SignalProcessing/RecoWire/DeconTools/BaselineMostProbAve_tool.cc
@@ -12,7 +12,7 @@
 #include "messagefacility/MessageLogger/MessageLogger.h"
 #include "cetlib_except/exception.h"
 #include "larcore/CoreUtils/ServiceUtil.h" // lar::providerFrom()
-#include "icaruscode/TPC/Utilities/SignalShapingICARUSService_service.h"
+#include "icaruscode/TPC/Utilities/SignalShapingICARUSService.h"
 #include "icarus_signal_processing/WaveformTools.h"
 
 #include <fstream>
