+++ icaruscode/TPC/SignalProcessing/RecoWire/DeconTools/BaselineStandard_tool.cc
@@ -10,7 +10,7 @@
 #include "art_root_io/TFileService.h"
 #include "messagefacility/MessageLogger/MessageLogger.h"
 #include "cetlib_except/exception.h"
-#include "icaruscode/TPC/Utilities/SignalShapingICARUSService_service.h"
+#include "icaruscode/TPC/Utilities/SignalShapingICARUSService.h"
 #include "larcore/CoreUtils/ServiceUtil.h" // lar::providerFrom()
 
 #include "TH1D.h"
