+++ icaruscode/TPC/SignalProcessing/RecoWire/DeconTools/FullWireDeconvolution_tool.cc
@@ -13,7 +13,7 @@
 #include "lardata/DetectorInfoServices/DetectorPropertiesService.h"
 #include "larcore/Geometry/Geometry.h"
 #include "larcore/CoreUtils/ServiceUtil.h" // lar::providerFrom()
-#include "icaruscode/TPC/Utilities/SignalShapingICARUSService_service.h"
+#include "icaruscode/TPC/Utilities/SignalShapingICARUSService.h"
 
 #include "art/Utilities/make_tool.h"
 #include "icarus_signal_processing/WaveformTools.h"
