+++ icaruscode/TPC/SignalProcessing/RecoWire/DeconTools/ROIDeconvolution_tool.cc
@@ -13,7 +13,7 @@
 #include "lardata/DetectorInfoServices/DetectorPropertiesService.h"
 #include "larcore/Geometry/Geometry.h"
 #include "larcore/CoreUtils/ServiceUtil.h" // lar::providerFrom()
-#include "icaruscode/TPC/Utilities/SignalShapingICARUSService_service.h"
+#include "icaruscode/TPC/Utilities/SignalShapingICARUSService.h"
 
 #include "icaruscode/TPC/SignalProcessing/RecoWire/DeconTools/IBaseline.h"
 #include "icarus_signal_processing/Filters/ICARUSFFT.h"
