+++ icaruscode/TPC/SignalProcessing/RecoWire/DeconTools/ROIFinderStandard_tool.cc
@@ -12,7 +12,7 @@
 #include "lardata/DetectorInfoServices/DetectorPropertiesService.h"
 #include "larcore/Geometry/WireReadout.h"
 #include "larcore/CoreUtils/ServiceUtil.h" // lar::providerFrom()
-#include "icaruscode/TPC/Utilities/SignalShapingICARUSService_service.h"
+#include "icaruscode/TPC/Utilities/SignalShapingICARUSService.h"
 
 #include "TH1D.h"
