+++ icaruscode/TPC/SignalProcessing/RecoWire/RecoWireROI_module.cc
@@ -45,7 +45,7 @@
 #include "lardata/DetectorInfoServices/DetectorPropertiesService.h"
 #include "lardata/ArtDataHelper/WireCreator.h"
 #include "lardata/Utilities/AssociationUtil.h"
-#include "icaruscode/TPC/Utilities/SignalShapingICARUSService_service.h"
+#include "icaruscode/TPC/Utilities/SignalShapingICARUSService.h"
 #include "icarus_signal_processing/Filters/ICARUSFFT.h"
 
 #include "larevt/CalibrationDBI/Interface/DetPedestalService.h"
