+++ icaruscode/TPC/Utilities/CMakeLists.txt
@@ -47,7 +47,10 @@ art_make_library(
 )
 
 cet_build_plugin(FileCatalogMetadataICARUS art::service LIBRARIES ${icarus_util_lib_list})
-cet_build_plugin(SignalShapingICARUSService art::service LIBRARIES ${icarus_util_lib_list})
+cet_build_plugin(SignalShapingICARUSService art::service
+	LIBRARIES PUBLIC icaruscode::TPC_Utilities # includes the service header
+	LIBRARIES PRIVATE ${icarus_util_lib_list}
+	)
 cet_build_plugin(SIOVChannelStatusICARUSService art::service LIBRARIES ${icarus_util_lib_list})
 cet_build_plugin(TFileMetadataICARUS art::service LIBRARIES ${icarus_util_lib_list})
